refactor(probability): separate services of lsst and ztf in probability api
