+++ multisurveys-apis/src/probability_api/services/ztf_service.py
@@ -1,0 +1,83 @@
+def taxonomy_data(taxonomy_list):
+    """
+    dict example: {classifier_name:{class_name:[],classifier_version:''}}
+    """
+
+    taxonomy_dict = {}
+
+    for n in range(len(taxonomy_list)):
+        taxonomy_dict[taxonomy_list[n].classifier_name] = taxonomy_list[n].__dict__
+
+    return taxonomy_dict
+
+
+def filter_data_by_higher_version(prob_dict):
+    """
+    Dictionary format: {'classifier_name': {'classifier_version: [{},{},{}]}}
+    """
+    data_by_higher_version = {}
+
+    for itemKey, itemValue in prob_dict.items():
+        if len(itemValue) > 1:
+            lastest_version = max(itemValue.keys())
+            if itemKey not in data_by_higher_version:
+                data_by_higher_version[itemKey] = {lastest_version: itemValue[lastest_version]}
+        else:
+            if itemKey not in data_by_higher_version:
+                data_by_higher_version[itemKey] = itemValue
+
+    return data_by_higher_version
+
+
+def group_data_by_classifier_dict(prob_lis):
+    group_data_by_classifier = {}
+
+    for item in prob_lis:
+        aux_dict = {
+            "classifier_name": item.classifier_name,
+            "classifier_version": item.classifier_version,
+            "class_name": item.class_name,
+            "probability": item.probability,
+            "ranking": item.ranking,
+        }
+        classifier_name = item.classifier_name
+        classifier_version = item.classifier_version
+
+        if item.classifier_name not in group_data_by_classifier:
+            group_data_by_classifier[classifier_name] = {}
+
+        if classifier_version not in group_data_by_classifier[classifier_name]:
+            group_data_by_classifier[classifier_name][classifier_version] = []
+
+        group_data_by_classifier[classifier_name][classifier_version].append(aux_dict)
+
+    return group_data_by_classifier
+
+
+def classifiers_options(group_prob_by_version):
+    class_dict = []
+    priorities = {
+        "lc_classifier": 0,
+        "lc_classifier_top": 1,
+        "lc_classifier_BHRF_forced_phot": 2,
+        "stamp_classifier": 3,
+        "LC_classifier_ATAT_forced_phot(beta)": 4,
+        "stamp_classifier_2025_beta": 5,
+    }
+
+    for key, value in priorities.items():
+        if key in group_prob_by_version:
+            class_dict.append({key: format_classifiers_name(key)})
+
+    return class_dict
+
+
+def format_classifiers_name(classifier_name):
+    classifier_name = classifier_name.replace("_", " ").title()
+
+    if classifier_name.find("Atat") != -1:
+        classifier_name = classifier_name.replace("Atat", "ATAT")
+    if classifier_name.find("Bhrf") != -1:
+        classifier_name = classifier_name.replace("Bhrf", "BHRF")
+
+    return classifier_name
